Implement Global Protected Middleware
which responds to any request to Endpoints with a path that is
prefixed (or contains in general) by /api/restricted, with an
error if there is not a current session with a valid username.
Otherwise, it invokes next().

Additionally, add restricted version of READ All Users Endpoint
to demonstrate protected middleware.